+++ CHANGELOG.md
@@ -130,6 +130,7 @@ high state of flux, you're at risk of it changing without notice.
     - deprecate `getMeetSemigroup`, use `min` instead (@gcanti)
     - deprecate `getJoinSemigroup`, use `max` instead (@gcanti)
     - deprecate `getDualSemigroup`, use `reverse` instead (@gcanti)
+    - deprecate `getStructSemigroup`, use `struct` instead (@gcanti)
   - `Show`
     - deprecate `showBoolean`, use `boolean.Show` instead (@gcanti)
     - deprecate `showString`, use `string.Show` instead (@gcanti)
@@ -341,6 +342,7 @@ high state of flux, you're at risk of it changing without notice.
     - add `intercalate` (@gcanti)
     - add `min` (@gcanti)
     - add `max` (@gcanti)
+    - add `struct` (@gcanti)
   - `State`
     - add `Pointed` instance (@gcanti)
     - add `Apply` instance (@gcanti)